Simplified Directions API to just one result type
