+++ src/main/java/com/google/maps/DirectionsApi.java
@@ -39,8 +39,8 @@ public class DirectionsApi {
   private DirectionsApi() {
   }
 
-  public static FullDirectionsApiRequest newRequest(GeoApiContext context) {
-    return new FullDirectionsApiRequest(context);
+  public static DirectionsApiRequest newRequest(GeoApiContext context) {
+    return new DirectionsApiRequest(context);
   }
 
   public static DirectionsApiRequest getDirections(GeoApiContext context,
@@ -49,31 +49,7 @@ public class DirectionsApi {
     return new DirectionsApiRequest(context).origin(origin).destination(destination);
   }
 
-  static class Response implements ApiResponse<DirectionsRoute[]> {
-    public String status;
-    public String errorMessage;
-    public DirectionsRoute[] routes;
-
-    @Override
-    public boolean successful() {
-      return "OK".equals(status) || "ZERO_RESULTS".equals(status);
-    }
-
-    @Override
-    public DirectionsRoute[] getResult() {
-      return routes;
-    }
-
-    @Override
-    public ApiException getError() {
-      if (successful()) {
-        return null;
-      }
-      return ApiException.from(status, errorMessage);
-    }
-  }
-
-  static class FullResponse implements ApiResponse<DirectionsResult> {
+  static class Response implements ApiResponse<DirectionsResult> {
     public String status;
     public String errorMessage;
     public GeocodedWaypoint[] geocodedWaypoints;
