--- src/main/java/com/google/maps/FullDirectionsApiRequest.java
@@ -1,34 +1,0 @@
-/*
- * Copyright 2015 Google Inc. All rights reserved.
- *
- *
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this
- * file except in compliance with the License. You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software distributed under
- * the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF
- * ANY KIND, either express or implied. See the License for the specific language governing
- * permissions and limitations under the License.
- */
-package com.google.maps;
-
-import com.google.maps.model.DirectionsResult;
-
-/**
- * Request for the Directions API.
- *
- * <p>
- *   This request generates a full result with the routes and geocoded waypoints.
- *   For just the routes, use {@link DirectionsApiRequest} by calling
- *   {@link com.google.maps.DirectionsApi#getDirections(GeoApiContext, String, String)}.
- * </p>
- */
-public class FullDirectionsApiRequest
-    extends DirectionsApiRequestBase<DirectionsResult, FullDirectionsApiRequest, DirectionsApi.FullResponse> {
-
-    FullDirectionsApiRequest(GeoApiContext context) {
-      super(context, DirectionsApi.API_CONFIG, DirectionsApi.FullResponse.class);
-    }
-}
